refactor: consolidate MemoEditor components - remove V2 suffixes
- Merge AttachmentListV2 + AttachmentItemCard into AttachmentList.tsx
- Merge RelationListV2 + RelationItemCard into RelationList.tsx
- Rename LocationDisplayV2 to LocationDisplay for consistency
- Update EditorMetadata to import new consolidated components
- Update component exports in index.ts
- Delete old separate component files

Benefits:
- 26% code reduction (379 → ~280 lines)
- Single source of truth per feature (list + card together)
- Clearer component hierarchy
- Better maintainability
