+++ web/src/components/MemoEditor/components/EditorMetadata.tsx
@@ -1,29 +1,26 @@
 import type { FC } from "react";
 import { useEditorContext } from "../state";
 import type { EditorMetadataProps } from "../types";
-import AttachmentListV2 from "./AttachmentListV2";
-import LocationDisplayV2 from "./LocationDisplayV2";
-import RelationListV2 from "./RelationListV2";
+import AttachmentList from "./AttachmentList";
+import LocationDisplay from "./LocationDisplay";
+import RelationList from "./RelationList";
 
 export const EditorMetadata: FC<EditorMetadataProps> = () => {
   const { state, actions, dispatch } = useEditorContext();
 
   return (
     <div className="w-full flex flex-col gap-2">
-      <AttachmentListV2
+      <AttachmentList
         attachments={state.metadata.attachments}
         localFiles={state.localFiles}
         onAttachmentsChange={(attachments) => dispatch(actions.setMetadata({ attachments }))}
         onRemoveLocalFile={(previewUrl) => dispatch(actions.removeLocalFile(previewUrl))}
       />
 
-      <RelationListV2
-        relations={state.metadata.relations}
-        onRelationsChange={(relations) => dispatch(actions.setMetadata({ relations }))}
-      />
+      <RelationList relations={state.metadata.relations} onRelationsChange={(relations) => dispatch(actions.setMetadata({ relations }))} />
 
       {state.metadata.location && (
-        <LocationDisplayV2 location={state.metadata.location} onRemove={() => dispatch(actions.setMetadata({ location: undefined }))} />
+        <LocationDisplay location={state.metadata.location} onRemove={() => dispatch(actions.setMetadata({ location: undefined }))} />
       )}
     </div>
   );
