+++ web/src/components/MemoEditor/components/index.ts
@@ -1,13 +1,11 @@
 // UI components for MemoEditor
 
-export { default as AttachmentItemCard } from "./AttachmentItemCard";
-export { default as AttachmentListV2 } from "./AttachmentListV2";
+export { default as AttachmentList } from "./AttachmentList";
 export * from "./EditorContent";
 export * from "./EditorMetadata";
 export * from "./EditorToolbar";
 export { FocusModeExitButton, FocusModeOverlay } from "./FocusModeOverlay";
 export { LinkMemoDialog } from "./LinkMemoDialog";
 export { LocationDialog } from "./LocationDialog";
-export { default as LocationDisplayV2 } from "./LocationDisplayV2";
-export { default as RelationItemCard } from "./RelationItemCard";
-export { default as RelationListV2 } from "./RelationListV2";
+export { default as LocationDisplay } from "./LocationDisplay";
+export { default as RelationList } from "./RelationList";
